[inspect] Simple find-one method
