+++ scat/src/main/java/scat/adapter/persistence/CountryJpaRepository.java
@@ -6,7 +6,6 @@ import scat.adapter.persistence.search.CountrySearch;
 import scat.domain.model.Country;
 import scat.domain.repo.CountryRepository;
 
-import javax.persistence.EntityNotFoundException;
 import java.util.Collection;
 import java.util.List;
 
@@ -18,11 +17,6 @@ public interface CountryJpaRepository extends JpaRepository<Country, Integer>, C
     @Override
     @Query("SELECT c FROM Country c WHERE LOWER(c.name) in (?1)")
     List<Country> findAllByNames(Collection<String> names);
-
-    @Override
-    default Country findOne(Integer id) {
-        return findById(id).orElseThrow(EntityNotFoundException::new);
-    }
 
     @Override
     default List<Country> findBy(Country params) {
