+++ scat/src/main/java/scat/adapter/persistence/SchoolJpaRepository.java
@@ -3,23 +3,17 @@ package scat.adapter.persistence;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import scat.adapter.persistence.search.SchoolSearch;
+import scat.adapter.persistence.support.SpecificationBuilder;
 import scat.domain.model.City;
 import scat.domain.model.School;
 import scat.domain.repo.SchoolRepository;
-import scat.adapter.persistence.support.SpecificationBuilder;
 
-import javax.persistence.EntityNotFoundException;
 import java.util.List;
 
 /**
  * @author levry
  */
 public interface SchoolJpaRepository extends JpaRepository<School, Long>, JpaSpecificationExecutor<School>, SchoolRepository {
-
-    @Override
-    default School findOne(Long id) {
-        return findById(id).orElseThrow(EntityNotFoundException::new);
-    }
 
     @Override
     default boolean hasSchool(String name, City city) {
