+++ scat/src/main/java/scat/domain/repo/SchoolTypeRepository.java
@@ -2,8 +2,9 @@ package scat.domain.repo;
 
 import scat.domain.model.SchoolType;
 
-import java.util.Arrays;
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
+import java.util.Optional;
 
 /**
  * @author levry
@@ -11,13 +12,15 @@ import java.util.List;
 public interface SchoolTypeRepository {
     List<SchoolType> findBy(SchoolType params);
 
-    SchoolType findOne(Integer id);
+    Optional<SchoolType> findById(Integer id);
 
     SchoolType save(SchoolType type);
-
-    SchoolType getOne(Integer id);
 
     void deleteById(Integer id);
 
     List<SchoolType> findAll();
+
+    default SchoolType findOne(Integer id) {
+        return findById(id).orElseThrow(EntityNotFoundException::new);
+    }
 }
